Complete CreatePurchase functionality
Removed CreatePurchase route from App.js because it is being called in
ProductIndex. ProductIndex is passing user, msgAlert, and product props
to CreatePurchase. CreatePurchase is a button that appears on
ProductIndex that does the POST request for a purchase with product's
id. Data in purchase.js axios call now sends the product id as an object